+++ demo/DemoV3/Model/Base/Address.generated.cs
@@ -29,7 +29,7 @@ namespace DemoV3.Model.Base {
         public virtual Dictionary<string, Flags> Flags { get; set; } = new Dictionary<string, Flags>();
 
         [JsonExtensionData]
-        public Dictionary<string, JToken> AdditionalProperties;
+        public Dictionary<string, JObject> AdditionalProperties;
 
         public Address ToAddress() {
             var item = new Address();
@@ -47,7 +47,7 @@ namespace DemoV3.Model.Base {
             City = other.City;
             Country = other.Country;
             Flags = other.Flags;
-            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, JToken>(other.AdditionalProperties) : null;
+            AdditionalProperties = other.AdditionalProperties != null ? new Dictionary<string, JObject>(other.AdditionalProperties) : null;
         }
 
         /// <summary>